+++ docs/changes.rst
@@ -7,6 +7,9 @@ Changelog
 - Mailing lists: Language specific lists for countries
   (`#1621 <https://github.com/syslabcom/scrum/issues/1621>`_)
   [reinhardt]
+- Mailing lists: Allow searching for path. Show path in result.
+  (`#1638 <https://github.com/syslabcom/scrum/issues/1638>`_)
+  [reinhardt]
 
 
 9.0.5 (2024-01-08)